requirements.txt
aiolimiter==1.1.0
beartype==0.12.0
beautifulsoup4==4.12.3
bottle==0.12.25
Farama_Notifications==0.0.4
Flask==3.0.2
graphviz==0.20.1
gymnasium==0.29.0
matplotlib==3.7.2
nltk==3.8.1
numpy==1.25.0
openai==0.27.0
pandas==1.5.3
Pillow==10.2.0
playwright
pytest==8.1.1
pytest_asyncio==0.23.6
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
setuptools==69.2.0
text_generation==0.7.0
tiktoken==0.4.0
tqdm==4.65.0
transformers==4.33.2
